Move utility method for getting the correct group for View Mod Files button to the new view model
