+++ src/NexusMods.App.UI/Controls/Trees/LoadoutItemGroupFileTreeViewModel.cs
@@ -113,47 +113,7 @@ public class LoadoutItemGroupFileTreeViewModel : AViewModel<IFileTreeViewModel>,
         }));
     }
     
-    /// <summary>
-    /// Returns the appropriate LoadoutItemGroup of files if the selection contains a LoadoutItemGroup containing files,
-    /// if the selection contains multiple LoadoutItemGroups of files, returns None.
-    /// </summary>
-    internal static Optional<LoadoutItemGroup.ReadOnly> GetViewModFilesLoadoutItemGroup(
-        IReadOnlyCollection<LoadoutItemId> loadoutItemIds, 
-        IConnection connection)
-    {
-        var db = connection.Db;
-        // Only allow when selecting a single item, or an item with a single child
-        if (loadoutItemIds.Count != 1) return Optional<LoadoutItemGroup.ReadOnly>.None;
-        var currentGroupId = loadoutItemIds.First();
-        
-        var groupDatoms = db.Datoms(LoadoutItemGroup.Group, Null.Instance);
 
-        while (true)
-        {
-            var childDatoms = db.Datoms(LoadoutItem.ParentId, currentGroupId);
-            if (childDatoms.Count == 0) return Optional<LoadoutItemGroup.ReadOnly>.None;
-
-            var childGroups = groupDatoms.MergeByEntityId(childDatoms);
-
-            // We have no child groups, check if children are files
-            if (childGroups.Count == 0)
-            {
-                return LoadoutItemWithTargetPath.TryGet(db, childDatoms[0].E, out _) 
-                    ? LoadoutItemGroup.Load(db, currentGroupId)
-                    : Optional<LoadoutItemGroup.ReadOnly>.None;
-            }
-            
-            // Single child group, check if that group is valid
-            if (childGroups.Count == 1)
-            {
-                currentGroupId = childGroups.First();
-                continue;
-            }
-        
-            // We have multiple child groups, return None
-            if (childGroups.Count > 1) return Optional<LoadoutItemGroup.ReadOnly>.None;
-        }
-    }
 
     private static FileTreeNodeViewModel CreateFolderNode(DirectoryData directory)
     {
